refactor(user): move all user DTOs to user/DTOs dir
